pghr13 refactoring
