+++ src/libsnark_helpers/libsnark_helpers.hpp
@@ -22,12 +22,11 @@
 
 // Contains required interfaces and types (keypair, proof, generator, prover, verifier)
 #include "zeth.h"
-#include <libsnark/zk_proof_systems/ppzksnark/r1cs_gg_ppzksnark/r1cs_gg_ppzksnark.hpp>
 #include <libsnark/gadgetlib1/gadget.hpp>
 #include "extended_proof.hpp"
 
 #ifdef LIBZETH_DEFAULT_SNARK_DEFINED
-#include "pghr13_response.hpp"
+#include "snarks/pghr13/pghr13_response.hpp"
 #include "snarks/pghr13/pghr13_computation.hpp"
 #elif SNARK_R1CS_GG_PPZKSNARK
 #include "groth16_response.hpp"//TODO
